Integration test reorg
