+++ myceli/src/listener.rs
@@ -130,165 +130,3 @@ impl Listener {
         Ok(())
     }
 }
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-
-    use anyhow::bail;
-    use assert_fs::fixture::ChildPath;
-    use assert_fs::{fixture::FileWriteBin, fixture::PathChild, TempDir};
-    use rand::{thread_rng, RngCore};
-    use std::thread;
-    use tokio::time::sleep;
-
-    struct TestListener {
-        listen_addr: String,
-        db_dir: TempDir,
-    }
-
-    impl TestListener {
-        pub fn new(listen_addr: &str) -> TestListener {
-            let db_dir = TempDir::new().unwrap();
-
-            return TestListener {
-                listen_addr: listen_addr.to_string(),
-                db_dir,
-            };
-        }
-
-        pub async fn start(&self) -> Result<()> {
-            let thread_listen_addr = self.listen_addr.to_owned();
-            let thread_db_path = self.db_dir.child("storage.db");
-
-            thread::spawn(move || start_listener_thread(thread_listen_addr, thread_db_path));
-
-            // A little wait so the listener can get listening
-            sleep(Duration::from_millis(50)).await;
-            Ok(())
-        }
-
-        pub fn generate_file(&self) -> Result<String> {
-            let mut data = Vec::<u8>::new();
-            data.resize(256, 1);
-            thread_rng().fill_bytes(&mut data);
-
-            let tmp_file = self.db_dir.child("test.file");
-            tmp_file.write_binary(&data)?;
-            Ok(tmp_file.path().to_str().unwrap().to_owned())
-        }
-    }
-
-    #[tokio::main]
-    async fn start_listener_thread(listen_addr: String, db_path: ChildPath) {
-        let db_path = db_path.path().to_str().unwrap();
-        let mut listener = Listener::new(&listen_addr, &db_path).await.unwrap();
-        listener
-            .listen()
-            .await
-            .expect("Error encountered in listener");
-    }
-
-    struct TestController {
-        socket: UdpSocket,
-        chunker: SimpleChunker,
-    }
-
-    impl TestController {
-        pub async fn new() -> Self {
-            let socket = UdpSocket::bind("127.0.0.1:0").await.unwrap();
-            let chunker = SimpleChunker::new(60);
-            TestController { socket, chunker }
-        }
-
-        pub async fn send_and_recv(
-            &mut self,
-            target_addr: &str,
-            message: Message,
-        ) -> Result<Message> {
-            self.send_msg(message, target_addr).await?;
-            self.recv_msg().await
-        }
-
-        pub async fn send_msg(&self, message: Message, target_addr: &str) -> Result<()> {
-            for chunk in self.chunker.chunk(message).unwrap() {
-                self.socket.send_to(&chunk, target_addr).await.unwrap();
-            }
-            Ok(())
-        }
-
-        pub async fn recv_msg(&mut self) -> Result<Message> {
-            let mut tries = 0;
-            loop {
-                let mut buf = vec![0; 128];
-                if self.socket.try_recv_from(&mut buf).is_ok() {
-                    match self.chunker.unchunk(&buf) {
-                        Ok(Some(msg)) => return Ok(msg),
-                        Ok(None) => {}
-                        Err(e) => bail!("Error found {e:?}"),
-                    }
-                }
-                tokio::time::sleep(Duration::from_millis(10)).await;
-                tries += 1;
-                if tries > 5 {
-                    bail!("Listen tries exceeded");
-                }
-            }
-        }
-    }
-
-    #[tokio::test]
-    pub async fn test_verify_listener_alive() {
-        let listener = TestListener::new("127.0.0.1:8080");
-        listener.start().await.unwrap();
-
-        let mut controller = TestController::new().await;
-
-        let response = controller
-            .send_and_recv(&listener.listen_addr, Message::request_available_blocks())
-            .await
-            .unwrap();
-
-        assert_eq!(response, Message::available_blocks(vec![]));
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    pub async fn test_transmit_receive_block() {
-        let transmitter = TestListener::new("127.0.0.1:8080");
-        let receiver = TestListener::new("127.0.0.1:8081");
-        let mut controller = TestController::new().await;
-
-        transmitter.start().await.unwrap();
-        receiver.start().await.unwrap();
-
-        let test_file_path = transmitter.generate_file().unwrap();
-        let resp = controller
-            .send_and_recv(
-                &transmitter.listen_addr,
-                Message::import_file(&test_file_path),
-            )
-            .await
-            .unwrap();
-        let root_cid = match resp {
-            Message::ApplicationAPI(ApplicationAPI::FileImported { cid, .. }) => cid,
-            other => panic!("Failed to receive FileImported msg {other:?}"),
-        };
-
-        controller
-            .send_msg(
-                Message::transmit_block(&root_cid, &receiver.listen_addr),
-                &transmitter.listen_addr,
-            )
-            .await
-            .unwrap();
-
-        sleep(Duration::from_millis(100)).await;
-
-        let resp = controller
-            .send_and_recv(&receiver.listen_addr, Message::request_available_blocks())
-            .await
-            .unwrap();
-
-        assert_eq!(resp, Message::available_blocks(vec![root_cid]));
-    }
-}
