+++ myceli/src/main.rs
@@ -1,11 +1,6 @@
-mod handlers;
-mod listener;
-mod receiver;
-mod transmit;
-
 use anyhow::Result;
 use clap::Parser;
-use listener::Listener;
+use myceli::listener::Listener;
 use tracing::Level;
 
 #[derive(Parser, Debug)]
